candidate-rag-chatbot/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
weaviate-client[agents]==4.11.1
gunicorn==21.2.0
